# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts API",
    "description": "API для управления контактами (CRUD), авторизация, email и фото.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    { "url": "http://localhost:3000" },
    { "url": "https://nodejs-hw-mongodb-rnrd.onrender.com" }
  ],
  "tags": [
    { "name": "Contacts", "description": "Operations with contacts" }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get all contacts",
        "operationId": "getContacts",
        "description": "Retrieve all contacts with optional filtering and pagination.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "example": 1 }, "description": "Page number" },
          { "in": "query", "name": "perPage", "schema": { "type": "integer", "example": 10 }, "description": "Contacts per page" },
          { "in": "query", "name": "name", "schema": { "type": "string", "example": "Dave" }, "description": "Filter by name" },
          { "in": "query", "name": "email", "schema": { "type": "string", "example": "dave@gmail.com" }, "description": "Filter by email" },
          { "in": "query", "name": "isFavourite", "schema": { "type": "boolean", "example": true }, "description": "Filter by favourite status" }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status","message","data"],
                  "properties": {
                    "status": { "type": "integer", "example": 200 },
                    "message": { "type": "string", "example": "Contacts retrieved successfully" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/contact" } }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contacts not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status","message"],
                  "properties": {
                    "status": { "type": "integer", "example": 404 },
                    "message": { "type": "string", "example": "Contacts not found" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Create a new contact",
        "operationId": "createContact",
        "description": "Create a new contact for the authenticated user.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name","phoneNumber","contactType"],
                "properties": {
                  "name": { "type": "string", "example": "John Deere", "description": "Contact name" },
                  "phoneNumber": { "type": "string", "example": "695642575512", "description": "Contact phone number" },
                  "email": { "type": "string", "example": "example@gmail.com", "description": "Contact email" },
                  "contactType": { "type": "string", "enum": ["work","home","personal"], "example": "personal", "description": "Contact type" },
                  "isFavourite": { "type": "boolean", "default": false, "example": true, "description": "Is favourite contact" },
                  "photo": { "type": "string", "format": "binary", "description": "Contact photo" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status","message","data"],
                  "properties": {
                    "status": { "type": "integer", "example": 201 },
                    "message": { "type": "string", "example": "Contact created successfully" },
                    "data": { "$ref": "#/components/schemas/contact" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request — invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status","message"],
                  "properties": {
                    "status": { "type": "integer", "example": 400 },
                    "message": { "type": "string", "example": "Invalid contact data provided" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get contact by ID",
        "operationId": "getContact",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string", "example": "68efd41636a23320bb4ef586" }, "description": "Contact ID" }
        ],
        "responses": {
          "200": {
            "description": "Contact retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status","message","data"],
                  "properties": {
                    "status": { "type": "integer", "example": 200 },
                    "message": { "type": "string", "example": "Contact retrieved successfully" },
                    "data": { "$ref": "#/components/schemas/contact" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status","message"],
                  "properties": {
                    "status": { "type": "integer", "example": 404 },
                    "message": { "type": "string", "example": "Contact not found" }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Contacts"],
        "summary": "Update contact",
        "operationId": "updateContact",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" }, "description": "Contact ID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "phoneNumber": { "type": "string" },
                  "email": { "type": "string" },
                  "contactType": { "type": "string", "enum": ["work","home","personal"] },
                  "isFavourite": { "type": "boolean" },
                  "photo": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Contact updated successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/contact" } } } },
          "404": { "description": "Contact not found" }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete contact",
        "operationId": "deleteContact",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" }, "description": "Contact ID" }
        ],
        "responses": {
          "200": { "description": "Contact deleted successfully" },
          "404": { "description": "Contact not found" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer" } },
    "schemas": {
      "contact": {
        "type": "object",
        "required": ["_id","name","phoneNumber","email","contactType","userId","createdAt","updatedAt"],
        "properties": {
          "_id": { "type": "string", "example": "68efd41636a23320bb4ef586" },
          "name": { "type": "string", "example": "John Deere" },
          "phoneNumber": { "type": "string", "example": "38284657512" },
          "email": { "type": "string", "example": "example@gmail.com" },
          "contactType": { "type": "string", "enum": ["work","home","personal"], "example": "personal" },
          "isFavourite": { "type": "boolean", "default": false, "example": true },
          "photo": { "type": "string", "format": "binary" },
          "userId": { "type": "string", "example": "68efd834f1781f960b1e7a6c" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}