# src/models/contact.js
import mongoose from "mongoose";

const contactSchema = new mongoose.Schema(
  {
    userId: { type: mongoose.Schema.Types.ObjectId, ref: "User", required: true },
    name: { type: String, required: true },
    phoneNumber: { type: String, required: true },
    email: { type: String },
    isFavourite: { type: Boolean, default: false },
    contactType: {
      type: String,
      enum: ["work", "home", "personal"],
      default: "personal",
      required: true,
    },
    photo: { type: String, default: null }
  },
  {
    timestamps: true,
    versionKey: false,
  }
);

export const Contact = mongoose.model("Contact", contactSchema);